+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu Sales
 
 * 0.8.5 (2016-09-30)
 
+    * BUGFIX      [OrderBundle]   Fixed simple account and simple contact widget to accept empty data.
     * BUGFIX      [OrderBundle]   Fixed sulu persistence for sales order.
 
 * 0.8.4 (2016-09-27)